+++ lib/ml_card_scanner.dart
@@ -1,5 +1,5 @@
 export 'src/model/card_info.dart';
-export 'src/model/scanner_exception.dart';
+export 'src/model/exceptions/scanner_exception.dart';
 export 'src/model/card_orientation.dart';
 export 'src/widget/scanner_widget.dart';
 export 'src/utils/scanner_widget_controller.dart';