+++ lib/src/model/scanning_params.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ml_card_scanner/src/model/card_info.dart';
-import 'package:ml_card_scanner/src/model/scanner_exception.dart';
+import 'package:ml_card_scanner/src/model/exceptions/scanner_exception.dart';
 
 
 class ScanningParams {