+++ lib/src/utils/card_parser_util.dart
@@ -17,11 +17,7 @@ class CardParserUtil {
   Future<CardInfo?> detectCardContent(InputImage inputImage) async {
     var input = await _textDetector.processImage(inputImage);
 
-    var clearElements = input.blocks
-        .map(
-          (e) => e.text.clean(),
-        )
-        .toList();
+    var clearElements = input.blocks.map((e) => e.text.clean()).toList();
 
     try {
       var possibleCardNumber = clearElements.firstWhere((input) {