+++ lib/src/widget/scanner_widget.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 import 'package:ml_card_scanner/ml_card_scanner.dart';
+import 'package:ml_card_scanner/src/model/typedefs.dart';
 import 'package:ml_card_scanner/src/utils/card_parser_util.dart';
 import 'package:ml_card_scanner/src/utils/logger.dart';
 import 'package:ml_card_scanner/src/widget/camera_overlay_widget.dart';
@@ -13,24 +14,25 @@ import 'package:ml_card_scanner/src/widget/text_overlay_widget.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 class ScannerWidget extends StatefulWidget {
-  final Widget? overlay;
   final CardOrientation overlayOrientation;
-  final Widget? overlayText;
+  final OverlayBuilder? overlayBuilder;
   final int scannerDelay;
   final bool oneShotScanning;
-
   final CameraResolution cameraResolution;
   final ScannerWidgetController? controller;
+  final CameraPreviewBuilder? cameraPreviewBuilder;
+  final OverlayTextBuilder? overlayTextBuilder;
 
   const ScannerWidget({
-    super.key,
-    this.overlay,
-    this.overlayText,
+    this.overlayBuilder,
     this.controller,
     this.scannerDelay = 400,
     this.oneShotScanning = true,
     this.overlayOrientation = CardOrientation.portrait,
     this.cameraResolution = CameraResolution.high,
+    this.cameraPreviewBuilder,
+    this.overlayTextBuilder,
+    super.key,
   });
 
   @override
@@ -41,21 +43,21 @@ class _ScannerWidgetState extends State<ScannerWidget>
     with WidgetsBindingObserver {
   final CardParserUtil _cardParser = CardParserUtil();
   final GlobalKey<CameraViewState> _cameraKey = GlobalKey();
+  final ValueNotifier<bool> _isCameraInitialized = ValueNotifier(false);
 
-  bool _isCameraInitialized = false;
   late CameraDescription _camera;
   late CameraController _cameraController;
   late ScannerWidgetController _scannerController;
 
   @override
   void initState() {
+    super.initState();
     if (mounted) {
       WidgetsBinding.instance.addObserver(this);
       _scannerController = widget.controller ?? ScannerWidgetController();
       _scannerController.addListener(_scanParamsListener);
       _initialize();
     }
-    super.initState();
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitDown,
       DeviceOrientation.portraitUp,
@@ -93,27 +95,36 @@ class _ScannerWidgetState extends State<ScannerWidget>
   Widget build(BuildContext context) {
     return Stack(
       children: <Widget>[
-        _isCameraInitialized
-            ? CameraWidget(
+        ValueListenableBuilder<bool>(
+          valueListenable: _isCameraInitialized,
+          builder: (context, cameraInitialized, _) {
+            if (cameraInitialized) {
+              return CameraWidget(
                 key: _cameraKey,
                 cameraController: _cameraController,
                 cameraDescription: _camera,
                 onImage: _detect,
                 scannerDelay: widget.scannerDelay,
-              )
-            : const SizedBox.shrink(),
-        widget.overlay ??
+                cameraPreviewBuilder: widget.cameraPreviewBuilder,
+              );
+            }
+
+            return const SizedBox.shrink();
+          },
+        ),
+        widget.overlayBuilder?.call(context) ??
             CameraOverlayWidget(
               cardOrientation: widget.overlayOrientation,
               overlayBorderRadius: 25,
               overlayColorFilter: Colors.black54,
             ),
-        Positioned(
-          left: 0,
-          right: 0,
-          bottom: (MediaQuery.of(context).size.height / 5),
-          child: widget.overlayText ?? const TextOverlayWidget(),
-        ),
+        widget.overlayTextBuilder?.call(context) ??
+            Positioned(
+              left: 0,
+              right: 0,
+              bottom: (MediaQuery.sizeOf(context).height / 5),
+              child: const TextOverlayWidget(),
+            ),
       ],
     );
   }
@@ -122,9 +133,7 @@ class _ScannerWidgetState extends State<ScannerWidget>
     try {
       var initializeResult = await _initializeCamera();
       if (initializeResult) {
-        setState(() {
-          _isCameraInitialized = initializeResult;
-        });
+        _isCameraInitialized.value = initializeResult;
       }
     } catch (e) {
       _handleError(ScannerException(e.toString()));
@@ -134,18 +143,18 @@ class _ScannerWidgetState extends State<ScannerWidget>
   Future<bool> _initializeCamera() async {
     var status = await Permission.camera.request();
     if (!status.isGranted) {
-      _handleError(ScannerException('Camera permission not granted.'));
+      _handleError(const ScannerPermissionIsNotGrantedException());
       return false;
     }
 
     final cameras = await availableCameras();
     if (cameras.isEmpty) {
-      _handleError(ScannerException('No cameras available.'));
+      _handleError(const ScannerNoCamerasAvailableException());
       return false;
     }
     if ((cameras.where((cam) => cam.lensDirection == CameraLensDirection.back))
         .isEmpty) {
-      _handleError(ScannerException('No back camera available.'));
+      _handleError(const ScannerNoBackCameraAvailableException());
       return false;
     }
     _camera = cameras
@@ -200,15 +209,11 @@ class _ScannerWidgetState extends State<ScannerWidget>
 
   void _handleData(CardInfo cardInfo) {
     final cardScannedCallback = _scannerController.onCardScanned;
-    if (cardScannedCallback != null) {
-      cardScannedCallback(cardInfo);
-    }
+    cardScannedCallback?.call(cardInfo);
   }
 
   void _handleError(ScannerException exception) {
     final errorCallback = _scannerController.onError;
-    if (errorCallback != null) {
-      errorCallback(exception);
-    }
+    errorCallback?.call(exception);
   }
 }